backend/requirements.txt
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
requests==2.31.0
google-generativeai==0.5.4
python-dotenv==1.0.1
